+++ src/app/meet/join/page.tsx
@@ -4,7 +4,7 @@ import TabBar from '@/components/TabBar';
 import Button from '@/components/Button';
 import MeetingJoinGuide from '@/features/meet/join/MeetingJoinGuide';
 import JoinCodeInput from '@/features/meet/join/JoinCodeInput';
-import { useInputState } from '@/hooks/useInputState';
+import { useInputState } from '@/lib/hooks/useInputState';
 import styled from 'styled-components';
 import { useRouter } from 'next/navigation';
 import Modal from '@/components/Modal';
