+++ src/app/signup/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import Logo from '/public/svgs/common/logo.svg';
 import React, { useState } from 'react';
-import { useInputState } from '@/hooks/useInputState';
+import { useInputState } from '@/lib/hooks/useInputState';
 import Modal from '@/components/Modal';
 import { useRouter } from 'next/navigation';
 import useToggle from '@/lib/hooks/useToggle';
