+++ src/features/account/EmailVerificationInputs/EmailVerificationInputs.tsx
@@ -1,7 +1,7 @@
 'use client';
 import styled from 'styled-components';
 import React from 'react';
-import { UseInputStateReturn } from '@/hooks/useInputState';
+import { UseInputStateReturn } from '@/lib/hooks/useInputState';
 import InputWrapper from '@/components/Input/InputWrapper';
 import Label from '@/components/common/Label';
 import Input from '@/components/Input';
