+++ src/features/account/signup/NicknameInput/NicknameInput.tsx
@@ -2,7 +2,7 @@
 import styled from 'styled-components';
 import React from 'react';
 import Button from '@/components/Button';
-import { UseInputStateReturn } from '@/hooks/useInputState';
+import { UseInputStateReturn } from '@/lib/hooks/useInputState';
 import InputWrapper from '@/components/Input/InputWrapper';
 import Label from '@/components/common/Label';
 import Input from '@/components/Input';
